+++ .gitignore
@@ -1,4 +1,3 @@
 # .gitignore
 keys/private.pem
 keys/public.pem
-node_modules
