+++ package.json
@@ -13,7 +13,6 @@
   "description": "",
   "dependencies": {
     "@types/bcrypt": "^5.0.2",
-    "@types/ioredis": "^4.28.10",
     "@types/jsonwebtoken": "^9.0.9",
     "@types/node": "^22.13.5",
     "bcrypt": "^5.1.1",
@@ -22,11 +21,9 @@
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
     "inversify": "^7.0.1",
-    "ioredis": "^5.6.0",
     "jsonwebtoken": "^9.0.2",
     "mongodb": "^6.13.1",
     "nodemon": "^3.1.9",
-    "password-validator-pro": "^1.1.2",
     "reflect-metadata": "^0.2.2",
     "ts-node": "^10.9.2",
     "typescript": "^5.7.3"
