+++ src/UserRepository.ts
@@ -1,32 +1,23 @@
 import { injectable, inject } from 'inversify';
 import { ObjectId } from 'mongodb';
 import { DependencyKeys } from './constant';
-import { IDatabase } from './mongo_connector/contracts/IDataBase';
-// import { UserRequest, UserResponse } from '../src/models/UserModel';
-import { UserRequest } from './models/userRequest';
-import { UserResponse } from './models/userResponse';
+import { IDatabase } from '../src/contracts/IDataBase';
+import { UserRequest, UserResponse, DatabaseError, DuplicateKeyError, NotFoundError } from '../src/models/UserModel';
 import { IUserRepository } from './contracts/IUserRepository';
-import { DatabaseError } from './errors/DBerror';
-import { DuplicateKeyError } from './errors/DuplicationError';
-import { NotFoundError } from './errors/NotFound';
-import { DBConfig } from './mongo_connector/DBConfigProvider';
 import bcrypt from 'bcrypt'
-import { IDBConfig } from './mongo_connector/contracts/IDBConfig';
 @injectable()
 export class UserRepository implements IUserRepository {
-  private db: IDatabase<any>;
+  private db: IDatabase;
   private readonly collectionName = 'users';
   private readonly saltRounds = 10;
-  private dbConfig: IDBConfig;
 
-  constructor(@inject(DependencyKeys.DatabaseAccess) db: IDatabase<any>, @inject(DependencyKeys.DBConfig) dbConfig:IDBConfig) {
+  constructor(@inject(DependencyKeys.DatabaseAccess) db: IDatabase) {
     this.db = db;
-    this.dbConfig=dbConfig
   }
 
   async createUserAsync(user: UserRequest): Promise<UserResponse> {
     try {
-      const existingUser = await this.db.findOne(this.collectionName, { email: user.email });
+      const existingUser = await this.db.findOne<UserRequest>(this.collectionName, { email: user.email });
       if (existingUser) {
         throw new DuplicateKeyError(`Duplicate email: ${user.email}`, { email: user.email });
       }
@@ -55,7 +46,7 @@ export class UserRepository implements IUserRepository {
 
   async getUserAsync(id: string): Promise<UserResponse | null> {
     try {
-      const user = await this.db.findOne(this.collectionName, { _id: new ObjectId(id) });
+      const user = await this.db.findOne<UserResponse>(this.collectionName, { _id: new ObjectId(id) });
       if (!user) return null;
       return new UserResponse(id, user.firstname, user.lastname, user.email, user.age, user.hashedPassword);
     } catch (error) {
@@ -68,12 +59,12 @@ export class UserRepository implements IUserRepository {
 
   async findUserAsync(email: string): Promise<UserResponse | null> {
     try {
-      const user = await this.db.findOne(this.collectionName, {email: email});
+      const user = await this.db.findOne<UserResponse>(this.collectionName, {email});
       if (!user){
-        throw new NotFoundError("Email not exist");
+        throw new NotFoundError("User Not exist");
       }
       if(!user._id){
-         throw new NotFoundError("Email user not exist");
+         throw new NotFoundError("User Not exist");
       }
       return new UserResponse(user._id.toString(), user.firstname, user.lastname, user.email, user.age, user.hashedPassword);
     } catch (error) {
@@ -90,16 +81,15 @@ export class UserRepository implements IUserRepository {
       const pageNum = Math.max(1, page);
       const limitNum = Math.max(1, limit);
       const skip = (pageNum - 1) * limitNum;
-      const total = await this.dbConfig.getCollection(this.collectionName).countDocuments();
+      const total = await this.db.getCollection<UserResponse>(this.collectionName).countDocuments();
 
       // Fetch the paginated users
-      const users : UserResponse[] = await this.dbConfig.getCollection<UserResponse>(this.collectionName)
+      const users = await this.db.getCollection<UserResponse>(this.collectionName)
         .find()
         .skip(skip)
         .limit(limitNum)
         .toArray();
-       
-      const data = users.map(user => new UserResponse((user._id as ObjectId).toString(), user.firstname, user.lastname, user.email, user.age, user.hashedPassword));
+      const data = users.map(user => new UserResponse(user._id.toString(), user.firstname, user.lastname, user.email, user.age, user.hashedPassword));
       const totalPages = Math.ceil(total / limitNum);
 
       return {
@@ -121,11 +111,11 @@ export class UserRepository implements IUserRepository {
 
   async updateUserAsync(id: string, user: UserRequest): Promise<boolean> {
     try {
-      const existingUser = await this.db.findOne(this.collectionName, { email: user.email });
+      const existingUser = await this.db.findOne<UserRequest &  { _id: ObjectId }>(this.collectionName, { email: user.email });
       if (existingUser && existingUser._id.toString() !== id) {
         throw new DuplicateKeyError(`Duplicate email: ${user.email}`, { email: user.email });
       }
-      return await this.db.updateOne(this.collectionName, { _id: new ObjectId(id) }, { $set: user });
+      return await this.db.updateOne<UserRequest>(this.collectionName, { _id: new ObjectId(id) }, { $set: user });
     } catch (error) {
       if (error instanceof Error) {
         if (error.message.includes('Duplicate')) {
@@ -146,29 +136,6 @@ export class UserRepository implements IUserRepository {
       }
       throw new DatabaseError(`Failed to delete user with ID ${id}`, 'Unknown error');
     }
-  }
-  async deleteUsersAsync(emails: string[]): Promise<{ deletedCount: number, notFoundEmails: string[] }> {
-      try{
-        // const notFoundEmails: string[] = [];
-        const existingUsers = await this.db.findAll(this.collectionName, { email: { $in: emails } });
-        console.log("Existing users", existingUsers)
-        const existingEmails = new Set(existingUsers.map(user => user.email));
-        const deletePromises = emails.map(async email=>{
-          if (!existingEmails.has(email)) 
-            return { email, deleted: false };
-        const deleted=await this.db.deleteOne(this.collectionName, { email });
-        return deleted?{email,deleted:true}:{email,deleted:false}
-        });
-        const results = await Promise.all(deletePromises);
-        const deletedCount = results.filter(result=>result.deleted).length;
-        const notFoundEmails=results.filter(result=>!result.deleted).map(result=>result.email);
-        return {deletedCount,notFoundEmails};
-      } catch(e){
-     if(e instanceof Error){
-      throw new DatabaseError(`Failed to delete user with email ${emails}`, e.message);
-     }
-     throw new DatabaseError(`Failed to delete user with email ${emails}`, 'Unknown error');
-      }
   }
 }
