--- src/contracts/IRedisClient.ts
@@ -1,5 +1,0 @@
-import Redis from "ioredis";
-export interface IRedisClient{
-    getclient():Redis;
-    disconnect():Promise<void>;
-}
