--- src/contracts/IRedisConfig.ts
@@ -1,7 +1,0 @@
-// import { Interface } from "readline";
-
-export interface IRediConfig{
-   host:string;
-   port:number;
-   password?:string
-}
