+++ src/contracts/IUserService.ts
@@ -7,5 +7,4 @@ export interface IUserService {
   getAllAsync(page: number, limit: number): Promise<{ data: UserResponse[], pagination: { total: number, page: number, limit: number, totalPages: number } }>;
   updateUserAsync(id: string, user: UserRequest): Promise<boolean>;
   deleteUserAsync(id: string): Promise<boolean>;
-  deleteUsersAsync(emails: string[]): Promise<{deletedCount: Number, notFoundEmails:string[]}>
 }
