+++ src/index.ts
@@ -1,37 +1,82 @@
+// import 'reflect-metadata';
+// import container from './container';
+// import { AppBuilder } from './AppBuilder';
+// import { DependencyKeys } from './constant';
+// import { DatabaseAccess } from './mongo_connector/DataBaseAccess';
+// import { UserRoutes } from './routes/UserRoutes';
+// require('dotenv').config();
+
+// process.on('unhandledRejection', (reason, promise) => {
+//   console.error('Unhandled Rejection at:', promise, 'Reason:', reason);
+// });
+
+// // Global handler for uncaught exceptions
+// process.on('uncaughtException', (error) => {
+//   console.error('Uncaught Exception:', error);
+// });
+
+// async function startServer() {
+//   const dbAccess = container.get<DatabaseAccess>(DependencyKeys.DatabaseAccess);
+//   const appBuilder = container.get<AppBuilder>(DependencyKeys.AppBuilder);
+//   const userRoute = container.get<UserRoutes>(DependencyKeys.Routes)
+
+//   try {
+//     await dbAccess.initialize();
+//   } catch (error) {
+//     console.error('Failed to initialize database connection:', error);
+//   }
+  
+//    // Parse the port from the environment variable, with a fallback to 3000
+//     const portEnv = process.env.PORT;
+
+//   const port :number = portEnv && !isNaN(parseInt(portEnv)) ? parseInt(portEnv) : 3000;
+//   appBuilder
+//   .withRequestLogger()
+//   .withCors({
+//     origin: '*',
+//     methods: ['GET', 'POST', 'PUT', 'DELETE'],
+//     allowedHeaders: ['Content-Type', 'Authorization'],
+//   })
+//   .withJsonContent()
+//   .withRoute('/', userRoute.getRouter())
+//   .withRouteHandler()
+//   .withErrorHandler()
+//   .build();
+
+// appBuilder.start(port);
+// }
+
+// startServer();
+
 import 'reflect-metadata';
 import container from './container';
 import { AppBuilder } from './AppBuilder';
 import { UserRoutes } from './routes/UserRoutes';
 import { DependencyKeys } from './constant';
-import { DatabaseAccess } from './mongo_connector/DBOperations';
-import { DBConfig } from './mongo_connector/DBConfigProvider';
-// import { RedisClient } from './redis/RedisClient';
-import { RedisClient } from './impl/RedisClient';
+import { DatabaseAccess } from './mongo_connector/DataBaseAccess';
 require('dotenv').config();
 
-process.on('unhandledRejection', (reason, promise) => {
-  console.error('Unhandled Rejection at:', promise, 'Reason:', reason);
-});
+// process.on('unhandledRejection', (reason, promise) => {
+//   console.error('Unhandled Rejection at:', promise, 'Reason:', reason);
+// });
 
-process.on('uncaughtException', (error) => {
-  console.error('Uncaught Exception:', error);
-});
+// process.on('uncaughtException', (error) => {
+//   console.error('Uncaught Exception:', error);
+// });
 
 const portEnv = process.env.PORT;
 const port: number = portEnv && !isNaN(parseInt(portEnv)) ? parseInt(portEnv) : 3000;
 
 async function startServer() {
-  const dbAccess = container.get<DBConfig>(DependencyKeys.DBConfig);
+  const dbAccess = container.get<DatabaseAccess>(DependencyKeys.DatabaseAccess);
   const appBuilder = container.get<AppBuilder>(DependencyKeys.AppBuilder);
   const userRoutes = container.get<UserRoutes>(DependencyKeys.Routes);
-  const redisClient = container.get<RedisClient>(DependencyKeys.RedisClient);
 
   try {
     await dbAccess.initialize();
-    console.log('Successfully Database initialised!');
-  } catch (e) {
-    console.error('Failure in initialisation of database:', e);
-    process.exit(1);
+    console.log('Database initialization completed successfully');
+  } catch (error) {
+    console.error('Failed to initialize database connection:', error);
   }
 
   appBuilder
@@ -42,23 +87,14 @@ async function startServer() {
       allowedHeaders: ['Content-Type', 'Authorization'],
     })
     .withApiContext('keys/private.pem', 'keys/public.pem')
-    .withPublicRoutes(['/register', '/login'])
+    .withPublicRoutes(['/api/register', '/api/login', '/api/public'])
     .withJsonContent()
     .withRoute('/', userRoutes.getRouter())
     .withRouteHandler()
     .withErrorHandler()
     .build();
 
-  const server = appBuilder.start(port);
-  const shutdown = async () => {
-    console.log('Trying to shut down server..');
-    await redisClient.disconnect();
-    console.log('Server closed');
-    process.exit(0);
-  };
-  
-  process.on('SIGINT', shutdown);
-  process.on('SIGTERM', shutdown);
+  appBuilder.start(port);
 }
 
 startServer();
